Avoid repeated materializing of entire schunk in subscriptor.
Instead, do a partial fill of the downloaded data, and deliver the
slice of it.  The result is equivalent to before, and still benefits
from the cache.
